Added UserDash, UserFeed, and ActivityCard
UserDash is the top level element for the User Dashboard page. This commit has the user feed implemented.
UserFeed is the user activity feed that makes up a list of ActivityCards. The list is basic right now, no styling done.
ActivityCard displays the name of the activity as long as the array of parameter key-value pairs. This is how the custom information will be stored.